Fix an issue with all Any annotations and early out detection
there was an early out docstrings that result in all `Any` (so that we don't put a comment on *everything*), but it did not properly take into account the need to update *existing* comments.  I changed the logic around to fix this
 
Also made a change to generate a type comment even if the docstrings result in all `Any` annotations: If the user provides type annotations we should honor them.